+++ src/app/core/index.ts
@@ -1,5 +1,5 @@
-export * from './auth-guard.service';
-export * from './can-deactivate-guard.service';
+export * from './auth.guard';
+export * from './can-deactivate.guard';
 export * from './config';
 export * from './entity.service';
 export * from './exception.service';
