+++ src/app/core/spinner/spinner.module.ts
@@ -1,7 +1,7 @@
 import { NgModule, Optional, SkipSelf } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
-import { throwIfAlreadyLoaded } from '../module-import-guard';
+import { throwIfAlreadyLoaded } from '../module-import-check';
 import { SpinnerComponent } from './spinner.component';
 import { SpinnerService } from './spinner.service';
