+++ src/app/core/toast/toast.module.ts
@@ -1,7 +1,7 @@
 import { NgModule, Optional, SkipSelf } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
-import { throwIfAlreadyLoaded } from '../module-import-guard';
+import { throwIfAlreadyLoaded } from '../module-import-check';
 import { ToastComponent } from './toast.component';
 import { ToastService } from './toast.service';
